Add cosmetic inventory locker and redesign store (#4962)
## Summary

- add a top-bar Inventory locker for all owned cosmetics, including
country flags
- add four-category loadout navigation and immediate card/swatch equip
behavior
- redesign Store cosmetics, effects, packs, subscriptions, affiliate
items, and custom currency around shared cosmetic cards
- keep Store selection and purchases isolated from equipped
`UserSettings`
- move purchase controls onto product tiles and fully remove the
cosmetic detail panel
- preserve ownership loading, retry, routing, checkout, and effect-slot
behavior

## Why

Owned cosmetics should be managed from a dedicated game-style inventory
instead of being equipped from the Store. Store remains the purchase
surface, while Inventory owns explicit equip actions.

## Card layout

Every surface — Inventory, Store cosmetics, effects, packs and
subscriptions — now renders the same `cosmetic-card`:

- The name leads the card, above the artwork. The equipped badge sits on
the artwork's bottom-left corner and the info bubble on its top-right,
so neither covers the title and names get the full card width. The info
bubble is anchored to a wrapper that starts below the name rather than
to the whole card, which keeps it outside the card's own button (a
button may not contain another control).
- Packs and subscriptions fit two cards per row on a phone, matching the
cosmetics and effects grids, and switch to a fixed width from `sm` up.
They previously kept one fixed-width card per row.
- Subscription previews size to their perk list instead of a square box,
which clipped the last perk at phone card widths.
- The owned tier's "Subscribed" status uses a purchase button's box —
full width, `min-h-11`, `text-base` — so it reads as a peer of the other
tiers' switch action.

## Naming colours

A pattern's palettes all share one name, so anything that named only the
pattern was ambiguous about which tile was acted on. The colour now
appears wherever an item is named back to the player:

- the selection toast: `Selected Ocean Stripes (Crimson)`
- the purchase confirmation: `Buy Ocean Stripes (Crimson) for 120
plutonium?`
- the insufficient-funds notice

Items with no palette are unchanged. Selecting a flag, crown, skin or
effect now raises the same toast — previously only patterns did. Unequip
stays silent, since it selects nothing.

## Validation

- `npm test` — 2,945 client/core and 298 server tests passed
- `npm run build-prod` — TypeScript and production Vite build passed
- ESLint, oxlint, Prettier, translation integrity, and `git diff
--check` passed
- driven in a headless browser at 1400x1000 and 390x844: grid wrapping,
badge and info placement, tooltip overflow, and the purchase
confirmation text were checked against the running app
- repository scan confirms no `CosmeticDetailPanel`,
`cosmetic-detail-panel`, or detail marker remains

Known existing warnings: Vitest emits a `MaxListenersExceededWarning`;
Vite reports large production chunks.

---------

+++ src/core/CosmeticSchemas.ts
@@ -145,7 +145,7 @@ export type TrailEffectType = (typeof TRAIL_EFFECT_TYPES)[number];
 //    onto the map — strands emerge from the unit, flare to full width, and
 //    spin with depth shading (facing segments bright, receding ones dark).
 //    `radius` = helix amplitude in tiles; `strands` = number of strands (the
-//    renderer clamps to 8); `rotationSpeed` = how fast the vortex spins, in
+//    renderer rounds and clamps to [1, 8]); `rotationSpeed` = how fast the vortex spins, in
 //    radians per second; the palette wraps once around the vortex
 //    circumference. radius must be positive (the geometry degenerates
 //    otherwise), so a non-positive value drops the entry like the enums. The
@@ -171,7 +171,7 @@ export const TrailEffectAttributesSchema = z.discriminatedUnion("type", [
     type: z.literal("spiral"),
     colors: z.array(z.string()),
     radius: z.number().positive(),
-    strands: z.number().int().positive(),
+    strands: z.number().positive(),
     rotationSpeed: z.number(),
   }),
 ]);
